[step4] refactor : Point 이동 로직을 Direction에 위임, Bar 객체 삭제
